+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## unreleased
 * BraintreeUIComponents
   * Add BraintreeUIComponents module
-  * Add `PayPalButton(color:width:action:)` to generate a PayPal-branded button for launching PayPal flow
+  * Add `PayPalButton(request:color:width:completion:)` to generate a PayPal-branded button for launching PayPal flow
   * Add `VenmoButton(authentication:universalLink:request:color:width:completion:)` to generate a Venmo-branded button for launching Venmo flow
+
 ## 7.1.0 (2025-11-20)
 * BraintreeThreeDSecure
   * Rename `AdditionalInformation` to `BTThreeDSecureAdditionalInformation`.